+++ intensio/core/obfuscation/intensio_replace.py
@@ -33,7 +33,7 @@ class Replace:
         self.pythonIncludeWords = "include/python/include_python_words.txt"
 
 
-    def EachLine(self, codeArg, eachLine, Dict):
+    def EachLine(self, codeArg, eachLine, Dict, forFilesName):
         getIndexLineList    = []
         returnLine          = []
         charValue           = []
@@ -103,21 +103,25 @@ class Replace:
                                                         if not re.match(r"\w|\\", getIndexLineList[indexExploreBefore]):
                                                             # -- Check if it's 'from' and 'import' word key in line to avoid replace name of file if variable is identic name to file -- #
                                                             getLine = "".join(getIndexLineList)
-                                                            if "import" in getLine:
-                                                                if "from" in getLine:
-                                                                    # -- Cut the line from the current index and check if it is not there is the keyword "import" in the line -- #
-                                                                    breakLine = getIndexLineList[:indexLine]
-                                                                    breakLine = "".join(breakLine)
-                                                                    if not "import" in breakLine:
-                                                                        # -- It's file because only 'from'key word -- #
-                                                                        checkCharAfterWord = 1
+                                                            if forFilesName == False:
+                                                                if "import" in getLine:
+                                                                    if "from" in getLine:
+                                                                        # -- Cut the line from the current index and check if it is not there is the keyword "import" in the line -- #
+                                                                        breakLine = getIndexLineList[:indexLine]
+                                                                        breakLine = "".join(breakLine)
+                                                                        if not "import" in breakLine:
+                                                                            # -- It's file because only 'from'key word -- #
+                                                                            checkCharAfterWord = 1
+                                                                        else:
+                                                                            checkCharAfterWord = 0
                                                                     else:
                                                                         checkCharAfterWord = 0
+                                                                # -- Check if after char find by 'regReplace' variable there no is ' or " -- #
+                                                                elif re.match(r"\"|\'", getIndexLineList[indexExploreAfter]):
+                                                                    checkCharAfterWord = 1
                                                                 else:
                                                                     checkCharAfterWord = 0
-                                                            # -- Check if after char find by 'regReplace' variable there no is ' or " -- #
-                                                            elif re.match(r"\"|\'", getIndexLineList[indexExploreAfter]):
-                                                                checkCharAfterWord = 1
+                                                            # -- Only for -rfn, --replacefilesname feature -- #
                                                             else:
                                                                 checkCharAfterWord = 0
                                                         else:
@@ -196,6 +200,9 @@ class Replace:
         countRecursFiles        = 0
 
         if codeArg == "python":
+            detectFiles = "py"
+            blockDir    = "__pycache__"
+
             functionsDefined        = r"def\s(\w+)"                             # Functions
             classDefined            = r"class\s(\w+)"                           # Classes
             variablesErrorDefined   = r"except(\s+\w+\sas\s)(\w)"               # Error variables
@@ -206,15 +213,12 @@ class Replace:
             quoteInRegex                        = r"\={1}\s*r[\"|\']{1}"    # If quote in regex
             quoteOfEndCommentariesMultipleLines = r"^\s*[\"|\']{3}\)?\.?"   # """ and ''' without before variables, if commentaries is over multiple lines and he finish by .format() funtion
             quoteIntoVariable                   = r".*\={1}\s*\w*\.?\w*[\(|\.]{1}[\"|\']{3}|.*\={1}\s*[\"|\']{3}" # """ and ''' with before variables
-            
-            detectFile  = "py"
-            blockDirs   = "__pycache__"
 
-        recursFiles = [f for f in glob.glob("{0}{1}**{1}*.{2}".format(outputArg, self.utils.Platform(), detectFile), recursive=True)]
+        recursFiles = [f for f in glob.glob("{0}{1}**{1}*.{2}".format(outputArg, self.utils.Platform(), detectFiles), recursive=True)]
 
-        # -- Find variables/classes/functions and mixed it -- #
+        # -- Replace variables/classes/functions to random strings with length defined -- #
         for file in recursFiles:
-            if blockDirs in file:
+            if blockDir in file:
                 continue
             else:
                 with open(file, "r") as readFile:
@@ -223,9 +227,7 @@ class Replace:
                         # -- Variables -- #     
                         search = re.search(variablesDefined, eachLine)
                         if search != None:
-                            # -- Detect mixer -- #
                             mixer = self.mixer.GetStringMixer(mixerLevelArg)
-                            # -- Add the mixer value of variables -- #
                             if search.group(1) not in variablesDict:
                                 variablesDict[search.group(1)] = mixer
                         
@@ -309,7 +311,7 @@ class Replace:
                 wordsIncludedNotFound.append(word)
 
         for file in recursFiles:
-            if blockDirs in file:
+            if blockDir in file:
                 continue
             else:
                 with open(file, "r") as readFile:
@@ -380,7 +382,7 @@ class Replace:
         with tqdm.tqdm(total=countRecursFiles) as pbar:
             for file in recursFiles:
                 pbar.update(1)
-                if blockDirs in file:
+                if blockDir in file:
                     continue
                 else:
                     with fileinput.input(file, inplace=True) as inputFile:
@@ -392,12 +394,12 @@ class Replace:
                                     # -- Check code into """ or ''' -- #
                                     if re.match(quoteIntoVariable, eachLine):
                                         checkQuotePassing += 1
-                                        eachLine = Replace.EachLine(self, codeArg, eachLine, allDict.items())
+                                        eachLine = Replace.EachLine(self, codeArg, eachLine, allDict.items(), False)
                                         print(eachLine)
                                         continue
                                     elif re.match(quoteOfCommentariesMultipleLines, eachLine) or re.match(quoteOfEndCommentariesMultipleLines, eachLine):
                                         checkQuotePassing += 1
-                                        eachLine = Replace.EachLine(self, codeArg, eachLine, allDict.items())
+                                        eachLine = Replace.EachLine(self, codeArg, eachLine, allDict.items(), False)
                                         print(eachLine)
                                         if checkQuotePassing == 2:
                                             checkQuotePassing = 0
@@ -410,13 +412,13 @@ class Replace:
                                         checkQuotePassing = 0
                                         continue
                                     else:
-                                        eachLine = Replace.EachLine(self, codeArg, eachLine, allDict.items())
+                                        eachLine = Replace.EachLine(self, codeArg, eachLine, allDict.items(), False)
                                         print(eachLine)
                                         continue
                             
         # -- Check if variables/classes/functions have been mixed -- #
         for file in recursFiles:
-            if blockDirs in file:
+            if blockDir in file:
                 continue
             else:
                 with open(file, "r") as readFile:
@@ -456,10 +458,10 @@ class Replace:
         remove              = Remove()
 
         if codeArg == "python":
-            detectFile          = "py"
-            blockDirs           = "__pycache__"
+            detectFiles         = "py"
+            blockDir           = "__pycache__"
 
-        recursFiles = [f for f in glob.glob("{0}{1}**{1}*.{2}".format(outputArg, utils.Platform(), detectFile), recursive=True)]
+        recursFiles = [f for f in glob.glob("{0}{1}**{1}*.{2}".format(outputArg, utils.Platform(), detectFiles), recursive=True)]
 
         for number in recursFiles:
                 countRecursFiles += 1
@@ -471,7 +473,7 @@ class Replace:
             for file in recursFiles:
                 checkPrint = 0 # initialize check print() func at the begining of each file
                 pbar.update(1)
-                if blockDirs in file:
+                if blockDir in file:
                     continue
                 else:
                     # -- Add a new first random line and move the old first line to the second line to avoid replacing it -- #
@@ -507,7 +509,7 @@ class Replace:
         for file in recursFiles: 
             with open(file, "r") as inputFile:
                 for eachLine in inputFile:
-                    if blockDirs in file:
+                    if blockDir in file:
                         continue
                     else:
                         if not eachLine:
@@ -523,3 +525,121 @@ class Replace:
                 return EXIT_FAILURE
         else:
             return EXIT_FAILURE
+
+
+    def FilesName(self, codeArg, outputArg, mixerLevelArg, verboseArg):
+        filesNameDict       = {}
+        filesNameDictNoExt  = {}
+        filesNameFound      = []
+        filesNameFoundNoExt = []
+        filesNameMixed      = []
+        countRecursFiles    = 0
+        checkRenameInDir    = 0
+        checkRenameInCode   = 0
+        unix                = False
+        win                 = False
+        currentPosition     = os.getcwd()
+        utils               = Utils()
+        mixer               = Mixer()
+        remove              = Remove()
+
+        if codeArg == "python":
+            detectFiles         = "py"
+            blockDir            = "__pycache__"
+            blockFile           = "__init__"
+    
+        recursFiles = [f for f in glob.glob("{0}{1}**{1}*.{2}".format(outputArg, utils.Platform(), detectFiles), recursive=True)]
+
+        for number in recursFiles:
+                countRecursFiles += 1
+
+        for file in recursFiles:
+            if blockDir in file or blockFile in file:
+                continue
+            else:
+                if "\"" in file:
+                    parseFilePath = file.split("\"")
+                    win = True
+                else:
+                    parseFilePath = file.split("/")
+                    unix = True
+
+                mixer = self.mixer.GetStringMixer(mixerLevelArg)
+                filesNameDict[parseFilePath[-1]] = mixer + ".py" 
+                filesNameDictNoExt[parseFilePath[-1].replace(".py", "")] = mixer
+                
+                filesNameFound.append(parseFilePath[-1])
+                filesNameMixed.append(mixer + ".py")
+
+                removeExt = parseFilePath[-1].replace(".py","")
+                filesNameFoundNoExt.append(removeExt)
+      
+            
+        # -- Diplay all files found with their mixed values if verbose arg is actived -- #
+        if verboseArg:
+            print("\n[+] Files name found with their mixed values :\n")
+            for key, value in filesNameDict.items():
+                print("-> {0} : {1}".format(key, value))
+        
+        print("\n[+] Running replace files name in {0} file(s)...\n".format(countRecursFiles))
+        
+        # -- Replace all files name to random strings with length defined -- #
+        with tqdm.tqdm(total=countRecursFiles) as pbar:
+            for file in recursFiles:
+                pbar.update(1)
+                if blockDir in file:
+                    continue
+                else:
+                    # -- Rename all files in python code -- #
+                    with fileinput.input(file, inplace=True) as inputFile:
+                        for eachLine in inputFile:
+                            try:
+                                for fileName in filesNameFound:
+                                   for fileNameNoExt in filesNameFoundNoExt:
+                                        if fileName in eachLine or fileNameNoExt in eachLine:
+                                            if not ".py" in eachLine:
+                                                if re.match(r"\s*from\s+|\s*import\s+", eachLine):
+                                                    eachLine = Replace.EachLine(self, codeArg, eachLine, filesNameDictNoExt.items(), True)
+                                                    print(eachLine)
+                                                    raise BreakLoop
+                                                else:
+                                                    continue
+                                            else: 
+                                                eachLine = Replace.EachLine(self, codeArg, eachLine, filesNameDict.items(), True)
+                                                print(eachLine)
+                                                raise BreakLoop
+                                        else:
+                                            continue
+                                print(eachLine) 
+                            except BreakLoop:
+                                continue
+
+                    # -- Rename all files in their directories -- #
+                    if "\"" in file:
+                        parseFilePath = file.split("\"")
+                        win = True
+                    else:
+                        parseFilePath = file.split("/")
+                        unix = True
+
+                    for key, value in filesNameDict.items():        
+                        if key == parseFilePath[-1]:
+                            parseFilePath.remove(parseFilePath[-1])
+                            if unix == True:
+                                parseFilePathToMove = "/".join(parseFilePath)
+                            else:
+                                parseFilePathToMove = "\"".join(parseFilePath)
+                            os.chdir(parseFilePathToMove) # Move in directory to rename python file
+                            os.rename(key, value)
+                        else:
+                            continue
+                    os.chdir(currentPosition) # Return to old position        
+
+        if (remove.Backslashes(codeArg, outputArg) == 0):    
+            return EXIT_SUCCESS        
+        else:
+            return EXIT_FAILURE
+
+
+
+