+++ intensio/test/python/basic/output/basicRAT-example/core/__init__.py
@@ -1,0 +1,3 @@
+TcbMoYIWpCeSIdaGVCenlGGCzCXqYajG="""
+"""
+exec(TcbMoYIWpCeSIdaGVCenlGGCzCXqYajG)